Fix+Test: marshalling int slices + decoder progress
